requirements.txt
fastapi[all]
python-jose[cryptography]
passlib[bcrypt] 
bcrypt==4.0.1
pymongo